Move ESP8266SdFat wrapper into it's own library
To be able to use as dependency in other libraries.
