+++ src/ps2/CMakeLists.txt
@@ -24,6 +24,7 @@ target_link_libraries(ps2_card PRIVATE
             hardware_flash
             hardware_dma
             mcfat
+            sd_fat
             psram)
